+++ applications/server/test/mock/mock.module.ts
@@ -1,0 +1,5 @@
+import { Mocker } from "@cheeket/koa-mock";
+
+class MockModule extends Mocker {}
+
+export default MockModule;
